In this commit, we see changes to the Devote Xcode data model.  This data model is the description of the Core Data persistence (ORM) model.
This commit also contains additions of Note Images meant to populate and provide description in the notes where this affair is documented.  The notes themselves are also part of this commit.  As an add, there is no difference to check.

The project file has references to the audio files.  However, the pictures for the notes don't seem to be recognized.