I added a function method to the View class.  This ensures that all Views inherit a hide keyboard function.  Also, this function utilizes the Swift to Obj-C bridge.
I also added a hide keyboard call in the add item in Content view.  Likewise, we empty the "task" state variable.  Upon making this state variable empty, the redraw of the textfield makes the textfield itself empty.

To be fair, this had me looking for the toggle keyboard back on.  When a user taps on the text field, I expect the keyboard to come on.   The simulator has this covered for now.

I also removed the Add button in the navigation bar (+).  This seems to make sense for now.  We will see how the lessons enable edit of items.